requirements.txt
python-dotenv
requests
scikit-learn
Flask
pyttsx3
SpeechRecognition
PyAudio
pytest
huggingface-hub
